requirements.txt
Flask>=1.0.0
google-cloud-firestore==1.1.0
google-cloud-storage==1.16.1
google-auth==1.6.3
gunicorn==19.9.0
six==1.11.0
